Crate tokio_macros

Macros for use with Tokio

Attribute Macros

main
Marks async function to be executed by selected runtime. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.

main_basic
Marks async function to be executed by selected runtime. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.

main_threaded
Marks async function to be executed by selected runtime. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.

test
Marks async function to be executed by runtime, suitable to test environment

test_basic
Marks async function to be executed by runtime, suitable to test environment

test_threaded
Marks async function to be executed by runtime, suitable to test environment