Crate tokio_macros

Macros for use with Tokio

Attribute Macros

main
Marks async function to be executed by the selected runtime. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.

main_fail
Always fails with the error message below.

main_rt
Marks async function to be executed by selected runtime. This macro helps set up a Runtime without requiring the user to use Runtime or Builder directly.

test
Marks async function to be executed by runtime, suitable to test environment

test_fail
Always fails with the error message below.

test_rt
Marks async function to be executed by runtime, suitable to test environment