tokio_macros

Attribute Macro main

#[main]

将 async 函数标记为由所选运行时执行。该宏帮助设置一个 Runtime， 无需用户直接使用 Runtime 或 Builder。

注意：该宏被设计为简单易用，面向不需要复杂配置的应用。如果所提供的功能不够用， 您可以考虑使用 Builder，它提供更强大的接口。

注意：该宏可用于任何函数，不仅仅是 main 函数。在非 main 函数上使用它 会让该函数表现得像同步函数一样（每次调用时启动一个新运行时）。如果该函数被频繁调用， 建议使用 runtime builder 创建运行时，以便跨调用复用。

Non-worker async function

注意：由该宏标记的 async 函数不会作为 worker 运行。预期行为是此函数内部派生其他任务。 在此处提供的函数中 await 其他 future 的性能不会像 worker 派生的任务那样快。

Runtime flavors

该宏可通过 flavor 参数配置不同的运行时配置。

Multi-threaded

要使用多线程运行时，可通过以下方式配置该宏：

#[tokio::main(flavor = "multi_thread", worker_threads = 10)]

The worker_threads option configures the number of worker threads, and defaults to the number of cpus on the system. This is the default flavor.

注意：多线程运行时需要 rt-multi-thread 特性标志。

Current-thread

要使用称为 current_thread 运行时的单线程运行时， 可通过以下方式配置该宏：

#[tokio::main(flavor = "current_thread")]

Local

To use the local runtime, the macro can be configured using

#[tokio::main(flavor = "local")]

Function arguments

除 main（特殊）外，任何函数都允许使用参数。

Usage

Set the name of the runtime

#[tokio::main(name = "my-runtime")]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_multi_thread()
        .enable_all()
        .name("my-runtime")
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Using the multi-threaded runtime

#[tokio::main]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_multi_thread()
        .enable_all()
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Using the current-thread runtime

The basic scheduler is single-threaded.

#[tokio::main(flavor = "current_thread")]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_current_thread()
        .enable_all()
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Using the local runtime

The local runtime is similar to the current-thread runtime but supports task::spawn_local.

#[tokio::main(flavor = "local")]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_current_thread()
        .enable_all()
        .build_local(tokio::runtime::LocalOptions::default())
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Set number of worker threads

#[tokio::main(worker_threads = 2)]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_multi_thread()
        .worker_threads(2)
        .enable_all()
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Configure the runtime to start with time paused

#[tokio::main(flavor = "current_thread", start_paused = true)]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

fn main() {
    tokio::runtime::Builder::new_current_thread()
        .enable_all()
        .start_paused(true)
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Note that start_paused requires the test-util feature to be enabled.

Rename package

use tokio as tokio1;

#[tokio1::main(crate = "tokio1")]
async fn main() {
    println!("Hello world");
}

Equivalent code not using #[tokio::main]

use tokio as tokio1;

fn main() {
    tokio1::runtime::Builder::new_multi_thread()
        .enable_all()
        .build()
        .unwrap()
        .block_on(async {
            println!("Hello world");
        })
}

Configure unhandled panic behavior

Available options are shutdown_runtime and ignore. For more details, see Builder::unhandled_panic.

此选项仅与 current_thread 运行时兼容。

#[cfg(tokio_unstable)]
#[tokio::main(flavor = "current_thread", unhandled_panic = "shutdown_runtime")]
async fn main() {
    let _ = tokio::spawn(async {
        panic!("This panic will shutdown the runtime.");
    }).await;
}

Equivalent code not using #[tokio::main]

#[cfg(tokio_unstable)]
fn main() {
    tokio::runtime::Builder::new_current_thread()
        .enable_all()
        .unhandled_panic(tokio::runtime::UnhandledPanic::ShutdownRuntime)
        .build()
        .unwrap()
        .block_on(async {
            let _ = tokio::spawn(async {
                panic!("This panic will shutdown the runtime.");
            }).await;
        })
}

Note: This option depends on Tokio’s unstable API. See the documentation on unstable features for details on how to enable Tokio’s unstable features.