Attribute Macro tokio_macros::test

#[test]

Marks async function to be executed by runtime, suitable to test environment

Usage

Multi-thread runtime

#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
async fn my_test() {
    assert!(true);
}

Using default

The default test runtime is single-threaded.

#[tokio::test]
async fn my_test() {
    assert!(true);
}

Configure the runtime to start with time paused

#[tokio::test(start_paused = true)]
async fn my_test() {
    assert!(true);
}

Note that start_paused requires the test-util feature to be enabled.

NOTE:

If you rename the Tokio crate in your dependencies this macro will not work. If you must rename the current version of Tokio because you’re also using an older version of Tokio, you must make the current version of Tokio available as tokio in the module where this macro is expanded.